Articles complete
I removed the 'measure-wide' class from each example so that we could
more clearly see the larger components for now until we create the
actual format for everything. We can add it back in if need be. I also
went and fixed any code which was showing a warning in the console from
React.